repo/cattura-dxf/requirements.txt
PyQt5
Pillow
ezdxf
